requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.8.0
soupsieve==2.7
sympy==1.14.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.52.2
typing_extensions==4.13.2
urllib3==2.4.0
